Monitor max/min for azimuth and elliptical angles
Note that:
azi=-90..90
eli=-45..45

+++ Form1.cs
@@ -15,23 +15,40 @@ namespace input
 {
     public partial class Form1 : Form
     {
-        //this.FormClosing += new FormClosingEventHandler(myForm_FormClosing);
-        //form.FormClosed += new Form.FormClosedEventHandler( myForm_FormClosing() );
 
         public Process subProcess;
 
         private string ARGUMENTS;
 
         private double st1, st2, st3, azi, ell, dop, pow, ms;
+        private int aziNorm, ellNorm, aziMax=-90, aziMin=90, ellMax=-45, ellMin=45;
+        private int counter = 0;
+        private int _x, _y, x_min, y_min, x_max, y_max, x_width, y_width;
+                      
+        public Form1()
+        {
+            InitializeComponent();
+            x_width = 500;
+            y_width = 500;
 
+            x_min = 800;
+            y_min = 100;
+
+            x_max = x_min + x_width;
+            y_max = y_min + y_width;
+
+            _x = (x_min+x_max)/2;
+            _y = (y_min + y_max) / 2;
+
+        }
 
         public void Form1_FormClosing(Object sender, FormClosingEventArgs e)
         {
-            if (sbProcess != null)
+            if (IsRunning(subProcess))
             {
                 try
                 {
-                    subProcess.Close();
+                    subProcess.Kill();
                     return;
                 }
                 catch (NullReferenceException)
@@ -48,9 +65,16 @@ namespace input
             //MessageBox.Show(messageBoxCS.ToString(), "FormClosed Event");
         }
 
-        public Form1()
+        public static bool IsRunning(Process process) //used to check if process is running
         {
-            InitializeComponent();
+            if (process != null)
+            {
+                try { Process.GetProcessById(process.Id); }
+                catch (InvalidOperationException) { return false; }
+                catch (ArgumentException) { return false; }
+                return true;
+            }
+            return false;
         }
 
         private void parseArgs()
@@ -74,8 +98,6 @@ namespace input
 
             textBoxArgs.Text = ARGUMENTS;
         }
-
-
 
         private void button1_Click(object sender, EventArgs e)
         {
@@ -135,6 +157,7 @@ namespace input
                 return;
             }
 
+            // copy measurements to variables
             st1 = Convert.ToDouble(mesurments[0]); // Stokes1
             st2 = Convert.ToDouble(mesurments[1]); // Stokes2
             st3 = Convert.ToDouble(mesurments[2]); // Stokes3
@@ -144,9 +167,61 @@ namespace input
             pow = Convert.ToDouble(mesurments[6]); // Power
             ms  = Convert.ToDouble(mesurments[7]); // Ms
 
-            // **************************************************
-            //  INSERT CODE HERE TO DO SOMETHING WITH THE VALUES
-            // **************************************************
+            //Monitor maximum and minimum angles
+            if (azi > aziMax)
+            {
+                aziMax = (int)azi;
+                textAziMax.Text = aziMax.ToString();
+            }
+
+            if (azi < aziMin)
+            {
+                aziMin = (int)azi;
+                textAziMin.Text = aziMin.ToString();
+            }
+
+            if (ell > ellMax)
+            {
+                ellMax = (int)ell;
+                textEllMax.Text = ellMax.ToString();
+            }
+
+            if (ell < ellMin)
+            {
+                ellMin = (int)ell;
+                textEllMin.Text = ellMin.ToString();
+            }
+            
+
+            //measurements run counter
+            counter++;
+            textBoxCounter.Text = counter.ToString();
+
+            // ***************************************
+            //   draw dot based on azi and ell values
+            // ***************************************
+ 
+            // azi and ell are between -90 and 90
+            //normalize to x_width and y_width (
+            aziNorm = Convert.ToInt32(azi) * x_width / 180 + x_width/2;
+            ellNorm = Convert.ToInt32(ell) * y_width / 90 + y_width/2;
+
+            //output variables
+            textAzi.Text = azi.ToString();
+            textEll.Text = ell.ToString();
+            textAziNorm.Text = aziNorm.ToString();
+            textEllNorm.Text = ellNorm.ToString();
+
+            //draw
+            //pictureBox1.Image = null;
+            //drawDot(aziNorm, ellNorm);
+
+            //update position
+            _x = x_min + aziNorm;
+            _y = y_min + ellNorm;
+            Invalidate();
+
+
 
             /*
              * A few interesting libraries:
@@ -159,8 +234,7 @@ namespace input
             // Debug
             textBoxOutput.AppendText(s + "\n");
         }
-
-
+        
         private void OutputDataReceivedEventHandler(object sender, DataReceivedEventArgs e)
         {
             if (!String.IsNullOrEmpty(e.Data))
@@ -178,30 +252,39 @@ namespace input
                     textBoxOutput.AppendText("Error: " + e.Data  + "\n");
                 }));
         }
-
-
+        
         private void ProcessExited(object sender, EventArgs e)
         {
-            this.BeginInvoke(new MethodInvoker(() =>
+            //if(IsRunning(subProcess))
             {
-                textBoxOutput.AppendText("Process exited\n");
-                subProcess.Close();
-                if (checkBoxRestart.Checked)
-                    StartProcessing();
+                this.BeginInvoke(new MethodInvoker(() =>
+                {
+                    textBoxOutput.AppendText("Process exited\n");
+                    subProcess.Close();
+                    if (checkBoxRestart.Checked)
+                        StartProcessing();
 
-            }));
+                }));
+            }
    
         }
 
         private void button1_Click_1(object sender, EventArgs e)
         {
-            try
+            if (IsRunning(subProcess))
             {
-                subProcess.Kill();
+                try
+                {
+                    subProcess.Kill();
+                }
+                catch (Exception ex)
+                {
+                    textBoxOutput.AppendText("Error: " + ex.Message + "\n");
+                }
             }
-            catch (Exception ex)
+            else
             {
-                textBoxOutput.AppendText("Error: " + ex.Message + "\n");
+                textBoxOutput.AppendText("Error: No process running\n");
             }
         }
 
@@ -242,32 +325,55 @@ namespace input
 
         private void textBoxFile_TextChanged(object sender, EventArgs e)
         {
-
-        }
-
-        private void label3_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label4_Click(object sender, EventArgs e)
-        {
-
+            parseArgs();
         }
 
         private void numericUpDown6_ValueChanged(object sender, EventArgs e)
         {
-
-        }
-
-        private void label9_Click(object sender, EventArgs e)
-        {
-
+            parseArgs();
         }
 
         private void portBox_ValueChanged(object sender, EventArgs e)
         {
+            parseArgs();
+        }
 
+        private void slotBox_ValueChanged(object sender, EventArgs e)
+        {
+            parseArgs();
+        }
+
+        private void wavelengthBox_ValueChanged(object sender, EventArgs e)
+        {
+            parseArgs();
+        }
+
+        private void commBox_ValueChanged(object sender, EventArgs e)
+        {
+            parseArgs();
+        }
+
+        private void bootBox_ValueChanged(object sender, EventArgs e)
+        {
+            parseArgs();
+        }
+
+        private void fileNameBox_TextChanged(object sender, EventArgs e)
+        {
+            parseArgs();
+        }
+
+        private void Form1_Paint(object sender, PaintEventArgs e)
+        {
+            e.Graphics.FillRectangle(Brushes.BlueViolet, _x, _y, 10, 10);
+            //e.Graphics.DrawImage(new Bitmap("mushroom.png"), _x, _y, 32, 32);
+        }
+
+        private void timer1_Tick(object sender, EventArgs e)
+        {
+            //_x += 1;
+            //_y += 1;
+            //Invalidate();
         }
